+++ src/lib/pages/execute/components/schema-execute/ExecuteBox.tsx
@@ -306,6 +306,7 @@ export const ExecuteBox = ({
               control={control}
               setValue={setValue}
               attachFundsOption={attachFundsOption}
+              labelBgColor="gray.900"
             />
           </GridItem>
           <GridItem>